+++ src/spec/consts.ts
@@ -88,6 +88,7 @@ export const Spec1dot2: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
   false,
   false,
   false,
+  false,
   true,
   false,
   false
@@ -149,6 +150,7 @@ export const Spec1dot3: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
   false,
   [],
   true,
+  false,
   false,
   true,
   true,
@@ -222,6 +224,7 @@ export const Spec1dot4: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
     VulnerabilityRatingMethod.Other
   ],
   true,
+  false,
   false,
   true,
   true,
@@ -324,6 +327,7 @@ export const Spec1dot5: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
     VulnerabilityRatingMethod.Other
   ],
   true,
+  false,
   true,
   true,
   true,
@@ -430,6 +434,7 @@ export const Spec1dot6: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
     VulnerabilityRatingMethod.SSVC,
     VulnerabilityRatingMethod.Other
   ],
+  true,
   true,
   true,
   true,
@@ -545,6 +550,7 @@ export const Spec1dot7: Readonly<_SpecProtocol> = Object.freeze(new _Spec(
     VulnerabilityRatingMethod.SSVC,
     VulnerabilityRatingMethod.Other
   ],
+  true,
   true,
   true,
   true,